+++ pkg/resource/events-logger-secret/grafana-agent-secret.go
@@ -35,10 +35,10 @@ func generateGrafanaAgentSecret(lc loggedcluster.Interface, credentialsSecret *v
 		ExtraSecret: extraSecret{
 			Name: fmt.Sprintf("%s-%s", clusterName, common.GrafanaAgentExtraSecretName()),
 			Data: map[string]string{
-				"logging-url":       fmt.Sprintf(common.LokiURLFormat, lokiURL),
-				"logging-tenant-id": lc.GetTenant(),
-				"logging-username":  writeUser,
-				"logging-password":  writePassword,
+				common.LoggingURL:      fmt.Sprintf(common.LokiURLFormat, lokiURL),
+				common.LoggingTenantID: lc.GetTenant(),
+				common.LoggingUsername: writeUser,
+				common.LoggingPassword: writePassword,
 			},
 		},
 	}